+++ MTManger/app/src/main/java/com/wonddak/mtmanger/ui/MainActivity.kt
@@ -1,155 +1,293 @@
-//package com.wonddak.mtmanger.ui
-//
-//import android.content.SharedPreferences
-//import android.os.Bundle
-//import android.util.Log
-//import android.view.Menu
-//import android.view.MenuItem
-//import android.view.View
-//import android.widget.Toast
-//import androidx.activity.viewModels
-//import androidx.appcompat.app.AppCompatActivity
-//import androidx.fragment.app.Fragment
-//import androidx.lifecycle.lifecycleScope
-//import androidx.viewpager.widget.ViewPager
-//import com.android.billingclient.api.*
-//import com.google.android.gms.ads.AdRequest
-//import com.google.android.gms.ads.AdView
-//import com.google.android.gms.ads.MobileAds
-//import com.google.android.material.bottomnavigation.BottomNavigationView
-//import com.wonddak.mtmanger.*
-//import com.wonddak.mtmanger.databinding.ActivityMainBinding
-//import com.wonddak.mtmanger.ui.fragments.*
-//import com.wonddak.mtmanger.viewModel.MTViewModel
-//import dagger.hilt.android.AndroidEntryPoint
-//import kotlinx.coroutines.flow.collect
-//import javax.inject.Inject
-//
-//@AndroidEntryPoint
-//class Main44Activity : AppCompatActivity() {
-//    lateinit var mAdView: AdView
-//
-//
-//    override fun onCreate(savedInstanceState: Bundle?) {
-//        super.onCreate(savedInstanceState)
-//        lifecycleScope.launchWhenCreated {
-//            mtViewModel.removeAdStatus.collect {
-//                Log.i("Billing JWH", "Main status $it")
-//                if (it) {
-//                    mAdView.visibility = View.GONE
-//                }
-//            }
-//        }
-//
-//        lifecycleScope.launchWhenCreated {
-//            mtViewModel.bottomMenuStatus.collect {
-//                if (it) {
-//                    mBottomNavigationView.visibility = View.VISIBLE
-//                } else {
-//                    mBottomNavigationView.visibility = View.GONE
-//                }
-//            }
-//        }
-//        lifecycleScope.launchWhenCreated {
-//            mtViewModel.topBackButtonStatus.collect {
-//                supportActionBar!!.setDisplayHomeAsUpEnabled(!it)
-//            }
-//        }
-//
-//        setSupportActionBar(binding.toolbarMain)
-//        supportActionBar!!.setDisplayShowTitleEnabled(false)
-//
-//        MobileAds.initialize(baseContext) {}
-//        mAdView = binding.adView
-//        val adRequest = AdRequest.Builder().build()
-//        mAdView.loadAd(adRequest)
-//
-//
-//        mBottomNavigationView = binding.bottomNavigation
-//        mViewPager = binding.ViewPager
-//        mViewPager.adapter = PagerAdapter(supportFragmentManager)
-//
-//        binding.ViewPager.addOnPageChangeListener(
-//            object : ViewPager.OnPageChangeListener {
-//                override fun onPageScrollStateChanged(state: Int) {}
-//
-//                override fun onPageScrolled(
-//                    position: Int,
-//                    positionOffset: Float,
-//                    positionOffsetPixels: Int
-//                ) {
-//                }
-//
-//                override fun onPageSelected(position: Int) {
-//                    // 네비게이션 메뉴 아이템 체크상태
-//                    mBottomNavigationView.menu.getItem(position).isChecked = true
-//                }
-//            })
-//
-//        initNavigationBar()
-//    }
-//
-//    override fun onBackPressed() {
-//        if (supportFragmentManager.backStackEntryCount == 0) {
-//
-//        } else if (supportFragmentManager.backStackEntryCount == 1) {
-//            supportFragmentManager.popBackStack()
-//            binding.mainactivitytitle.text = "MT 매니저"
-//            mtViewModel.setTopButtonStatus(false)
-//        } else {
-//            supportFragmentManager.popBackStack()
-//        }
-//
-//    }
-//
-//    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-//        val prefs: SharedPreferences = applicationContext.getSharedPreferences("mainMT", 0)
-//        val hide = prefs.getBoolean("hide", false)
-//        if (!hide) {
-//            menuInflater.inflate(R.menu.menu_main, menu)
-//        } else {
-//            menuInflater.inflate(R.menu.menu_main2, menu)
-//        }
-//        return true
-//    }
-//
-//
-//    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-//        when (item.itemId) {
-//            R.id.action_settings -> {
-//                if (supportFragmentManager.backStackEntryCount == 0) {
-//                    supportFragmentManager.beginTransaction()
-//                        .addToBackStack(null)
-//                        .add(R.id.settingfrag, SettingFragment())
-//                        .commit()
-//
-//
-//                    mBottomNavigationView.visibility = View.GONE
-//                    mtViewModel.setTopButtonStatus(false)
-//                    binding.mainactivitytitle.text = "설정"
-//                }
-//            }
-//            R.id.action_hide_bottom -> {
-//                mtViewModel.toggleBottomMenuStatus()
-//                invalidateOptionsMenu()
-//            }
-//            android.R.id.home -> {
-//                supportFragmentManager.popBackStack()
-//                val prefs: SharedPreferences = applicationContext.getSharedPreferences("mainMT", 0)
-//                val hide = prefs.getBoolean("hide", false)
-//                if (!hide) {
-//                    mBottomNavigationView.visibility = View.VISIBLE
-//                } else {
-//                    mBottomNavigationView.visibility = View.GONE
-//                }
-//                binding.mainactivitytitle.text = "MT 매니저"
-//                mtViewModel.setTopButtonStatus(true)
-//            }
-//
-//        }
-//        return super.onOptionsItemSelected(item)
-//    }
-//}
-//
-//
-//
+package com.wonddak.mtmanger.ui
+
+import android.content.SharedPreferences
+import android.os.Bundle
+import android.widget.Toast
+import androidx.activity.ComponentActivity
+import androidx.activity.OnBackPressedCallback
+import androidx.activity.compose.setContent
+import androidx.activity.viewModels
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.expandHorizontally
+import androidx.compose.animation.expandVertically
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.animation.shrinkHorizontally
+import androidx.compose.animation.shrinkVertically
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.NavigationBarItemDefaults
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
+import androidx.navigation.NavController
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
+import com.wonddak.mtmanger.BillingModule
+import com.wonddak.mtmanger.R
+import com.wonddak.mtmanger.model.BottomNavItem
+import com.wonddak.mtmanger.ui.theme.MTMangerTheme
+import com.wonddak.mtmanger.ui.theme.match1
+import com.wonddak.mtmanger.ui.theme.match2
+import com.wonddak.mtmanger.ui.view.BuyView
+import com.wonddak.mtmanger.ui.view.MainView
+import com.wonddak.mtmanger.ui.view.MtListView
+import com.wonddak.mtmanger.ui.view.PersonView
+import com.wonddak.mtmanger.ui.view.PlanView
+import com.wonddak.mtmanger.ui.view.SettingView
+import com.wonddak.mtmanger.ui.view.common.DefaultText
+import com.wonddak.mtmanger.ui.view.common.NoDataBase
+import com.wonddak.mtmanger.viewModel.MTViewModel
+import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.launch
+import javax.inject.Inject
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    private var backKeyPressedTime: Long = 0
+
+    private val mtViewModel: MTViewModel by viewModels()
+
+    @Inject
+    lateinit var billingModule: BillingModule
+
+    @Inject
+    lateinit var preferences: SharedPreferences
+
+    private val callback = object : OnBackPressedCallback(true) {
+        override fun handleOnBackPressed() {
+            if (System.currentTimeMillis() > backKeyPressedTime + 2000) {
+                backKeyPressedTime = System.currentTimeMillis();
+                Toast.makeText(this@MainActivity, "한번 더 누르시면 종료됩니다.", Toast.LENGTH_SHORT).show()
+                return
+            }
+            if (System.currentTimeMillis() <= backKeyPressedTime + 2000) {
+                finish()
+            }
+        }
+    }
+
+    @OptIn(ExperimentalMaterial3Api::class)
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        this.onBackPressedDispatcher.addCallback(this, callback)
+
+        val mainmtid: Int = preferences.getInt("id", 0)
+
+        mtViewModel.setMtId(mainmtid)
+
+        lifecycleScope.launch {
+            lifecycle.repeatOnLifecycle(Lifecycle.State.RESUMED) {
+                billingModule.removeAddStatus.collect {
+                    mtViewModel.setRemoveAddStatus(it)
+                }
+            }
+
+        }
+
+        setContent {
+            MTMangerTheme {
+                val navController = rememberNavController()
+
+                Scaffold(
+                    topBar = {
+                        TopAppBar(mtViewModel = mtViewModel)
+                    },
+                    bottomBar = {
+                        AnimatedVisibility(
+                            !mtViewModel.showSetting && !mtViewModel.showMtList,
+                            enter = expandVertically(),
+                            exit = shrinkVertically(),
+                        ) {
+                            BottomNavigationBar(navController = navController)
+                        }
+                    },
+                    containerColor = match1
+                ) {
+                    Box(Modifier.padding(it)) {
+                        AnimatedVisibility(
+                            !mtViewModel.showSetting && !mtViewModel.showMtList,
+                            enter = fadeIn(),
+                            exit = fadeOut()
+                        ) {
+                            NavGraph(
+                                navController = navController,
+                                mtViewModel
+                            )
+                        }
+                        AnimatedVisibility(
+                            mtViewModel.showMtList,
+                            enter = expandHorizontally(),
+                            exit = shrinkHorizontally()
+                        ) {
+                            MtListView(mtViewModel = mtViewModel)
+                        }
+                        AnimatedVisibility(
+                            mtViewModel.showSetting,
+                            enter = expandHorizontally(),
+                            exit = shrinkHorizontally()
+                        ) {
+                            SettingView(mtViewModel = mtViewModel, billingModule = billingModule)
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun NavGraph(
+    navController: NavHostController,
+    mtViewModel: MTViewModel
+) {
+
+    NavHost(navController = navController, startDestination = BottomNavItem.Main.screenRoute) {
+
+        composable(BottomNavItem.Main.screenRoute) {
+            NoDataBase(
+                mtViewModel
+            ) {
+                MainView(mtViewModel = mtViewModel)
+            }
+        }
+        composable(BottomNavItem.Person.screenRoute) {
+            NoDataBase(mtViewModel) {
+                PersonView(
+                    mtViewModel = mtViewModel
+                )
+            }
+        }
+        composable(BottomNavItem.Buy.screenRoute) {
+            NoDataBase(mtViewModel) {
+                BuyView(
+                    mtViewModel = mtViewModel
+                )
+            }
+        }
+        composable(BottomNavItem.Plan.screenRoute) {
+            NoDataBase(mtViewModel) {
+                PlanView(
+                    mtViewModel = mtViewModel
+                )
+            }
+        }
+
+    }
+}
+
+@Composable
+fun BottomNavigationBar(navController: NavController) {
+    val items = listOf(
+        BottomNavItem.Main,
+        BottomNavItem.Person,
+        BottomNavItem.Buy,
+        BottomNavItem.Plan,
+    )
+
+    NavigationBar(
+        containerColor = match2,
+    ) {
+        val navBackStackEntry by navController.currentBackStackEntryAsState()
+        val currentRoute = navBackStackEntry?.destination?.route
+        items.forEach { item ->
+            NavigationBarItem(
+                selected = currentRoute == item.screenRoute,
+                onClick = {
+                    navController.navigate(item.screenRoute) {
+                        navController.graph.startDestinationRoute?.let { screen_route ->
+                            popUpTo(screen_route) {
+                                saveState = true
+                            }
+                        }
+                        launchSingleTop = true
+                        restoreState = true
+                    }
+                },
+                icon = {
+                    Icon(
+                        painter = painterResource(id = item.icon),
+                        contentDescription = null
+                    )
+                },
+                label = {
+                    DefaultText(
+                        text = item.title,
+                        color = if (currentRoute == item.screenRoute) match1 else Color.White
+                    )
+                },
+                colors = NavigationBarItemDefaults.colors(
+                    selectedIconColor = match1,
+                    unselectedIconColor = Color.White,
+                    indicatorColor = match2
+                ),
+            )
+        }
+    }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun TopAppBar(
+    mtViewModel: MTViewModel
+) {
+    TopAppBar(
+        title = {
+            DefaultText(
+                text = "MT 매니저",
+                color = match1
+            )
+        },
+        colors = TopAppBarDefaults.smallTopAppBarColors(
+            containerColor = match2
+        ),
+        actions = {
+            AnimatedVisibility(visible = !mtViewModel.showMtList) {
+                IconButton(
+                    onClick = {
+                        mtViewModel.showSetting = true
+                    },
+                    enabled = !mtViewModel.showSetting
+                ) {
+                    Icon(
+                        painter = painterResource(id = R.drawable.ic_baseline_settings_24),
+                        contentDescription = null,
+                        tint = match1,
+                    )
+                }
+            }
+        },
+        navigationIcon = {
+            AnimatedVisibility(mtViewModel.showSetting || mtViewModel.showMtList) {
+                IconButton(
+                    onClick = {
+                        mtViewModel.showSetting = false
+                        mtViewModel.showMtList = false
+                    }
+                ) {
+                    Icon(
+                        painter = painterResource(id = R.drawable.back_arrow),
+                        contentDescription = null,
+                        tint = match1,
+                    )
+                }
+            }
+        }
+    )
+}
