--- MTManger/app/src/main/java/com/wonddak/mtmanger/ui/common/dialog/BaseDialog.kt
@@ -1,46 +1,0 @@
-package com.wonddak.mtmanger.ui.common.dialog
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.annotation.LayoutRes
-import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.DialogFragment
-import androidx.fragment.app.activityViewModels
-import androidx.viewbinding.ViewBinding
-import com.wonddak.mtmanger.viewModel.MTViewModel
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-
-abstract class BaseDialog<B:ViewBinding>(
-    @LayoutRes val idRes: Int
-) :DialogFragment() {
-    protected lateinit var binding : B
-
-    protected val mMainScope = CoroutineScope(Dispatchers.Main)
-    protected val mBgDispatcher: CoroutineDispatcher = Dispatchers.IO
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        isCancelable = false
-    }
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = DataBindingUtil.inflate(inflater,idRes,container,false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        initBinding()
-    }
-
-    abstract fun initBinding()
-}
