--- MTManger/app/src/main/java/com/wonddak/mtmanger/ui/common/fragment/BaseDataBindingFragment.kt
@@ -1,33 +1,0 @@
-package com.wonddak.mtmanger.ui.common.fragment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.annotation.LayoutRes
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-
-abstract class BaseDataBindingFragment <B : ViewDataBinding>
-    (@LayoutRes private val resId: Int) : CommonFragment() {
-
-    protected lateinit var binding: B
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = DataBindingUtil.inflate(inflater,resId,container,false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.lifecycleOwner = this
-        initBinding()
-    }
-
-    abstract fun initBinding()
-
-}
