+++ MTManger/app/src/main/java/com/wonddak/mtmanger/ui/view/MtListView.kt
@@ -2,6 +2,8 @@ package com.wonddak.mtmanger.ui.view
 
 import android.widget.Toast
 import androidx.activity.compose.BackHandler
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -11,18 +13,24 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Divider
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.wonddak.mtmanger.noRippleClickable
+import com.wonddak.mtmanger.room.MtData
 import com.wonddak.mtmanger.ui.theme.maple
+import com.wonddak.mtmanger.ui.theme.match1
 import com.wonddak.mtmanger.ui.theme.match2
+import com.wonddak.mtmanger.ui.view.common.DefaultText
 import com.wonddak.mtmanger.viewModel.MTViewModel
 
 @Composable
@@ -37,6 +45,7 @@ fun MtListView(
     Column(
         Modifier
             .fillMaxSize()
+            .background(match1)
             .padding(horizontal = 10.dp, vertical = 5.dp),
         verticalArrangement = Arrangement.SpaceBetween
     ) {
@@ -46,7 +55,6 @@ fun MtListView(
             Spacer(
                 modifier = Modifier
                     .height(10.dp)
-                    .padding(vertical = 5.dp)
             )
             Text(
                 text = "MT 리스트",
@@ -56,18 +64,53 @@ fun MtListView(
                 modifier = Modifier.fillMaxWidth(),
                 textAlign = TextAlign.Center
             )
-            LazyColumn() {
-                items(mtList) {
-                    Text(
-                        text = it.mtTitle,
-                        modifier = Modifier.noRippleClickable() {
-                            Toast.makeText(context, "${it.mtTitle}로 변경했어요", Toast.LENGTH_SHORT)
-                                .show()
-                            mtViewModel.setMtId(it.mtDataId!!)
-                        }
+            Spacer(
+                modifier = Modifier
+                    .height(10.dp)
+            )
+            LazyColumn(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .border(2.dp, match2, RoundedCornerShape(10.dp))
+                    .padding(5.dp)
+            ) {
+                items(mtList) { mtdata ->
+                    MtListItem(mtData = mtdata) {
+                        Toast.makeText(
+                            context,
+                            "${mtdata.mtTitle}로 변경했어요",
+                            Toast.LENGTH_SHORT
+                        )
+                            .show()
+                        mtViewModel.setMtId(mtdata.mtDataId!!)
+                        mtViewModel.showMtList = false
+                    }
+                    Divider(
+                        color = match2
                     )
                 }
             }
         }
+    }
+}
+
+@Composable
+fun MtListItem(
+    mtData: MtData,
+    onClick: () -> Unit
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .noRippleClickable(onClick)
+            .padding(10.dp),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        DefaultText(
+            text = mtData.mtTitle
+        )
+        DefaultText(
+            text = "${mtData.mtStart} ~ ${mtData.mtEnd}"
+        )
     }
 }
