+++ MTManger/app/src/main/java/com/wonddak/mtmanger/ui/view/dialog/DatePicker.kt
@@ -1,7 +1,8 @@
-package com.wonddak.mtmanger.ui.dialog
+package com.wonddak.mtmanger.ui.view.dialog
 
 import android.app.DatePickerDialog
 import android.content.Context
+import com.wonddak.mtmanger.R
 import java.util.Calendar
 
 object DatePicker {
@@ -15,7 +16,7 @@ object DatePicker {
         val day = calendar.get(Calendar.DAY_OF_MONTH)
 
         DatePickerDialog(
-            context, onDateSetListener, year, month, day
+            context, R.style.DatePicker, onDateSetListener, year, month, day
         ).show()
     }
 
@@ -31,7 +32,7 @@ object DatePicker {
         val day = calendar.get(Calendar.DAY_OF_MONTH)
 
         val picker = DatePickerDialog(
-            context, onDateSetListener, year, month, day
+            context, R.style.DatePicker, onDateSetListener, year, month, day
         )
         picker.datePicker.apply {
             this.minDate = minDate
@@ -51,7 +52,7 @@ object DatePicker {
         val day = calendar.get(Calendar.DAY_OF_MONTH)
 
         val picker = DatePickerDialog(
-            context, onDateSetListener, year, month, day
+            context, R.style.DatePicker, onDateSetListener, year, month, day
         )
         picker.datePicker.apply {
             this.minDate = minDate
