+++ MTManger/app/src/main/java/com/wonddak/mtmanger/ui/view/dialog/MtDialog.kt
@@ -17,7 +17,6 @@ import androidx.compose.ui.text.input.KeyboardType
 import com.wonddak.mtmanger.R
 import com.wonddak.mtmanger.noRippleClickable
 import com.wonddak.mtmanger.room.MtData
-import com.wonddak.mtmanger.ui.dialog.DatePicker
 import com.wonddak.mtmanger.ui.view.common.DialogBase
 import com.wonddak.mtmanger.ui.view.common.DialogTextField
