--- MTManger/app/src/main/res/drawable-v24/btn_layout2.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-
-    <stroke
-        android:width="3dp"
-        android:color="@color/match2"/>
-    <size
-        android:width="36dp"
-        android:height="36dp"/>
-    <corners android:radius="10dp"/>
-    <solid
-        android:color="@color/match1"/>
-
-
-</shape>
