--- MTManger/app/src/main/res/drawable-v24/solid_layout.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-
-    <stroke
-        android:width="1dp"
-        android:color="@color/match1"/>
-    <solid
-        android:color="@color/match2"/>
-
-
-</shape>
