--- MTManger/app/src/main/res/drawable/solid_layout2.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-
-    <stroke
-        android:width="3dp"
-        android:color="@color/match2"/>
-
-    <corners android:radius="5dp"/>
-    <solid
-        android:color="@color/match1"/>
-
-
-</shape>
