--- MTManger/app/src/main/res/menu/menu_main.xml
@@ -1,17 +1,0 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/action_hide_bottom"
-        android:icon="@drawable/ic_baseline_arrow_drop_down_24"
-        android:title="하단바 숨기기"
-        app:showAsAction="always" />
-
-    <item
-        android:id="@+id/action_settings"
-        android:icon="@drawable/ic_baseline_settings_24"
-        android:title="설정"
-        app:showAsAction="always" />
-
-
-</menu>
