--- MTManger/app/src/main/res/menu/menu_main2.xml
@@ -1,17 +1,0 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/action_hide_bottom"
-        android:icon="@drawable/ic_baseline_arrow_drop_up_24"
-        android:title="하단바 올리기"
-        app:showAsAction="always" />
-
-    <item
-        android:id="@+id/action_settings"
-        android:icon="@drawable/ic_baseline_settings_24"
-        android:title="설정"
-        app:showAsAction="always" />
-
-
-</menu>
