--- MTManger/app/src/main/res/menu/my_navigation_items.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/nav_main"
-        android:title="메인"
-        android:icon="@drawable/ic_baseline_home_24"/>
-    <item
-        android:id="@+id/nav_person"
-        android:title="참가자 명단"
-        android:icon="@drawable/ic_baseline_person_outline_24"/>
-    <item
-        android:id="@+id/nav_buy"
-        android:title="구매내역"
-        android:icon="@drawable/ic_baseline_attach_money_24"/>
-    <item
-        android:id="@+id/nav_plan"
-        android:title="계획"
-        android:icon="@drawable/ic_baseline_book_24"/>
-
-</menu>
